% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{box_3.7}
\alias{box_3.7}
\title{Box 3.7 from FEDAI}
\format{
A data.frame
}
\usage{
box_3.7
}
\description{
This hypothetical dataset has two columns, a realized treatment (\code{Z}) and a revealed outcome (\code{Y}). The purpose of the dataset is to illustrate how traditional t-tests and randomization inference can yield different answers
}
\keyword{datasets}
