% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clingingsmith_etal_2009}
\alias{clingingsmith_etal_2009}
\title{clingingsmith_etal_2009 from FEDAI}
\format{
A data.frame
}
\usage{
clingingsmith_etal_2009
}
\description{
David Clingingsmith, Asim Ijaz Khwaja, Michael Kremer,
Estimating the Impact of The Hajj:
Religion and Tolerance in Islam's Global Gathering,
The Quarterly Journal of Economics, Volume 124, Issue 3, August 2009, Pages 1133–1170,
https://doi.org/10.1162/qjec.2009.124.3.1133
}
\details{
These data describe 958 who were randomly selected to go on Hajj via lottery. The outcome Y_views describes these individuals' post-treatment views of others.
}
\keyword{datasets}
