% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{lm_bootstrap}
\alias{lm_bootstrap}
\title{Bootstrap linear model estimates using lm_robust}
\usage{
lm_bootstrap(formula, data, times = 1000, ...)
}
\arguments{
\item{formula}{A formula specifying the model.}

\item{data}{A data frame.}

\item{times}{Number of bootstrap replications. Default is 1000.}

\item{...}{Additional arguments passed to \code{estimatr::lm_robust}.}
}
\value{
A data frame with class \code{lm_bootstrap}, containing bootstrap estimates of model coefficients.
}
\description{
This function runs bootstrap resampling to estimate uncertainty around
regression coefficients using \code{estimatr::lm_robust} with \code{se_type = "none"}.
}
\examples{
boot_out <- lm_bootstrap(mpg ~ wt + hp, data = mtcars, times = 100)
broom::tidy(boot_out, alpha = 0.05)

}
