% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_variances.R
\name{pop_sd}
\alias{pop_sd}
\title{Population standard deviation}
\usage{
pop_sd(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{logical. Should missing values be removed?}
}
\description{
Divides by \code{N} rather than \code{N - 1}. See \code{\link[=pop_var]{pop_var()}}.
}
\examples{
x <- c(1, 2, 3, 4)
pop_sd(x) # divides by n

# For comparision
sd(x) # divides by n - 1
}
