% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_variances.R
\name{sd_var}
\alias{sd_var}
\title{Population standard deviation}
\usage{
sd_var(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{logical. Should missing values be removed?}
}
\description{
Population standard deviation
}
\examples{
x <- c(1, 2, 3, 4)
sd_var(x) # divides by n

# For comparision
sd(x) # divides by n - 1
}
