% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_2.1}
\alias{table_2.1}
\title{Table 2.1 from FEDAI}
\format{
A data.frame
}
\usage{
table_2.1
}
\description{
This hypothetical dataset shows the treated (\code{Y_Z_1}) and untreated (\code{Y_Z_0}) potential outcomes for all seven villages, as well as a column for their difference, the village-level treatment effect \code{tau_i}
}
\keyword{datasets}
