% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_2.2}
\alias{table_2.2}
\title{Table 2.2 from FEDAI}
\format{
A data.frame
}
\usage{
table_2.2
}
\description{
This hypothetical dataset shows a realized assignment (\code{Z}) and the corresponding revealed outcome (\code{Y}). Unobserved treated (\code{Y_Z_1}) and untreated (\code{Y_Z_0}) potential outcomes are marked as \code{NA} as is the village-level treatment effect \code{tau_i}.
}
\keyword{datasets}
