% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_3.7}
\alias{table_3.7}
\title{Table 3.7 from FEDAI}
\format{
A data.frame
}
\usage{
table_3.7
}
\description{
This hypothetical dataset has the treated (\code{Y_Z_1}) and untreated (\code{Y_Z_0}) potential outcomes for 12 classrooms in 4 schools (low sampling variability)
}
\keyword{datasets}
