% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_4.1}
\alias{table_4.1}
\title{Table 4.1 from FEDAI}
\format{
A data.frame
}
\usage{
table_4.1
}
\description{
Schedule of potential outcomes, pretest scores, and group assignments for simulated teacher incentives experiment
}
\keyword{datasets}
