% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_4.2}
\alias{table_4.2}
\title{Table 4.2 from FEDAI}
\format{
A data.frame
}
\usage{
table_4.2
}
\description{
Block random assignment of teacher incentives experiment
}
\keyword{datasets}
