% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_6.4}
\alias{table_6.4}
\title{Table 6.4 from FEDAI}
\format{
A data.frame
}
\usage{
table_6.4
}
\description{
Hypothetical schedule of potential outcomes and assigned treatments for illustration of a downstream experiment
}
\keyword{datasets}
