% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_7.2}
\alias{table_7.2}
\title{Table 7.2 from FEDAI}
\format{
A data.frame
}
\usage{
table_7.2
}
\description{
Complete set of potential outcomes and covariates for eight subjects
}
\keyword{datasets}
