% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arceneaux_2005}
\alias{arceneaux_2005}
\title{arceneaux_2005 from FEDAI}
\format{
A data.frame
}
\usage{
arceneaux_2005
}
\description{
Arceneaux, K. (2005). Using Cluster Randomized Field Experiments to Study Voting Behavior. The ANNALS of the American Academy of Political and Social Science, 601(1), 169-179.
Precinct-level cluster-randomized field experiment in Kansas City, Missouri. Columns: \code{precinct}, \code{Z} (treatment assignment), \code{D_contact} (whether contact with the household was successful), \code{Y_voted_2003} (outcome: voted in 2003 election).
}
\keyword{datasets}
