% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arceneaux_2005}
\alias{arceneaux_2005}
\title{arceneaux_2005 from FEDAI}
\format{
A data.frame
}
\usage{
arceneaux_2005
}
\description{
Arceneaux, K. (2005). Using Cluster Randomized Field Experiments to Study Voting Behavior. The ANNALS of the American Academy of Political and Social Science, 601(1), 169-179
These data are from a precinct-level randomized field experiment conducted in Kansas City, Missouri. The columns in this subset of the data include the precinct, whether the subject voted in the 2003 election (the outcome variable), whether contact was successful, and the assigned treatment status)
}
\keyword{datasets}
