% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bhavnani_2009}
\alias{bhavnani_2009}
\title{bhavnani_2009 from FEDAI}
\format{
A data.frame
}
\usage{
bhavnani_2009
}
\description{
Bhavnani, R. R. (2009). Exercise dataset for FEDAI Chapter 10 (Exercise 10.4). Study of gender quotas and voter turnout in India.
Columns: \code{controltreat} (control/treatment indicator), \code{genderofwinner} (gender of winning candidate), \code{turnout} (voter turnout), \code{turnout97} (1997 baseline turnout).
}
\keyword{datasets}
