% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation_3.4.R
\name{equation_3.4}
\alias{equation_3.4}
\title{The true standard error of a two-arm experiment under complete random assignment}
\usage{
equation_3.4(var_Y0, var_Y1, cov_Y0_Y1, N, m)
}
\arguments{
\item{var_Y0}{the variance of the untreated potential outcomes}

\item{var_Y1}{the variance of the treated potential outcomes}

\item{cov_Y0_Y1}{the covariance of the treated and untreated potential outcomes}

\item{N}{the number of units in the experiment}

\item{m}{the number of units assigned to treatment}
}
\description{
The true standard error of a two-arm experiment under complete random assignment
}
\examples{
var_Y0 <- pop_var(table_2.1$Y_Z_0)
var_Y1 <- pop_var(table_2.1$Y_Z_1)
cov_Y0_Y1 <- pop_cov(table_2.1$Y_Z_0, table_2.1$Y_Z_1)

equation_3.4(var_Y0 = var_Y0,
             var_Y1 = var_Y1,
             cov_Y0_Y1 = cov_Y0_Y1,
             N = 7,
             m = 2)

}
