% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_variances.R
\name{pop_cov}
\alias{pop_cov}
\title{Population covariance}
\usage{
pop_cov(x, y, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}

\item{na.rm}{logical. Should missing values be removed?}
}
\description{
Population covariance
}
\examples{
x <- c(1, 2, 3, 4)
y <- c(4, 3, 2, 1)
pop_cov(x, y)
# for comparison
cov(x, y)
}
