% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_variances.R
\name{pop_var}
\alias{pop_var}
\title{Population variance}
\usage{
pop_var(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{logical. Should missing values be removed?}
}
\description{
Population variance
}
\examples{
x <- c(1, 2, 3, 4)
pop_var(x) # divides by n

# For comparision
var(x) # divides by n - 1
}
