% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rosen_2010}
\alias{rosen_2010}
\title{rosen_2010 from FEDAI}
\format{
A data.frame
}
\usage{
rosen_2010
}
\description{
Jeff Rosen
The Effects of Race and Grammar Quality on the Responsiveness of American State Legislators: A Field Experiment,
Working paper, 2010.
}
\details{
These data describe 400 observations of state legislators who were sent letters requesting information about immigration rules. The race of the sender was randomly manipulated through the name (Colin or Jose), as was the grammar of the letter (good or bad). The dataset records whether the request received a response and whether the legislator is themselves hispanic or not.
}
\keyword{datasets}
