% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rosen_2010}
\alias{rosen_2010}
\title{rosen_2010 from FEDAI}
\format{
A data.frame
}
\usage{
rosen_2010
}
\description{
Jeff Rosen. The Effects of Race and Grammar Quality on the Responsiveness of American State Legislators: A Field Experiment. Working paper, 2010.
}
\details{
400 letters sent to state legislators in a 2x2 factorial experiment varying the apparent ethnicity of the author and the quality of the grammar. Columns: \code{observation}, \code{letter_name} (name used on letter), \code{X_legislator_hispanic} (1 = legislator is Hispanic), \code{Z_hispanic} (1 = Hispanic name), \code{Z_grammar_label} (factor: "good"/"bad"), \code{Z_good_grammar} (1 = good grammar), \code{Y_response} (1 = legislator replied).
}
\keyword{datasets}
