% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_10_1}
\alias{table_10_1}
\title{Table 10.1 from FEDAI}
\format{
A data.frame
}
\usage{
table_10_1
}
\description{
Hypothetical potential outcomes for six kinds of subjects
}
\keyword{datasets}
