% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_12_5}
\alias{table_12_5}
\title{Table 12.5 from FEDAI}
\format{
A data.frame
}
\usage{
table_12_5
}
\description{
Hypothetical potential outcomes for the Chong, De La O, Karlan, and Wantchekon (2011) municipal audit information experiment. Columns: type (honest or corrupt municipality), Y_Z_0 (turnout under control), Y_Z_1 (turnout under treatment).
}
\keyword{datasets}
