% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_3_4}
\alias{table_3_4}
\title{Table 3.4 from FEDAI}
\format{
A data.frame
}
\usage{
table_3_4
}
\description{
Hypothetical schedule of potential outcomes for 14 villages under noncompliance. Columns: \code{Village}, \code{Block}, \code{D} (dose received), \code{Y} (observed outcome).
}
\keyword{datasets}
