% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_3_5}
\alias{table_3_5}
\title{Table 3.5 from FEDAI}
\format{
A data.frame
}
\usage{
table_3_5
}
\description{
Hypothetical schedule of potential outcomes for 14 villages under noncompliance (no Village ID column). Columns: \code{Block}, \code{D} (dose received), \code{Y} (observed outcome).
}
\keyword{datasets}
