% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_3_6}
\alias{table_3_6}
\title{Table 3.6 from FEDAI}
\format{
A data.frame
}
\usage{
table_3_6
}
\description{
Hypothetical potential outcomes for 12 classrooms in 4 schools (high sampling variability). Columns: \code{school}, \code{classroom}, \code{Y_Z_0}, \code{Y_Z_1}, \code{Y_cluster_mean_Z_0}, \code{Y_cluster_mean_Z_1}.
}
\keyword{datasets}
