% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_4_1}
\alias{table_4_1}
\title{Table 4.1 from FEDAI}
\format{
A data.frame
}
\usage{
table_4_1
}
\description{
Schedule of potential outcomes and pretest scores for a simulated teacher incentives experiment. Columns: \code{observation}, \code{X_pretest} (pretest score, strong covariate), \code{X_pretest_weak} (pretest score, weak covariate), \code{Z} (realized treatment assignment), \code{Y_Z_0}, \code{Y_Z_1}.
}
\keyword{datasets}
