% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_4_2}
\alias{table_4_2}
\title{Table 4.2 from FEDAI}
\format{
A data.frame
}
\usage{
table_4_2
}
\description{
Block random assignment of teacher incentives experiment. Columns: \code{observation}, \code{block}, \code{X_pretest}, \code{X_pretest_weak}, \code{Z} (realized treatment assignment), \code{prob} (assignment probability), \code{wt} (inverse-probability weight), \code{Y_Z_0}, \code{Y_Z_1}.
}
\keyword{datasets}
