% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_5_1}
\alias{table_5_1}
\title{Table 5.1 from FEDAI}
\format{
A data.frame
}
\usage{
table_5_1
}
\description{
Hypothetical schedule of potential outcomes assuming one-sided noncompliance
}
\keyword{datasets}
