% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_6_1}
\alias{table_6_1}
\title{Table 6.1 from FEDAI}
\format{
A data.frame
}
\usage{
table_6_1
}
\description{
Hypothetical schedule of potential outcomes assuming two-sided noncompliance
}
\keyword{datasets}
