% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_7_7}
\alias{table_7_7}
\title{Table 7.7 from FEDAI}
\format{
A data.frame
}
\usage{
table_7_7
}
\description{
Hypothetical paired potential outcomes with attrition for four subjects. Columns: \code{observation}, \code{Pair}, \code{R_Z_0} (response indicator under control), \code{R_Z_1} (response indicator under treatment), \code{Y_Z_0}, \code{Y_Z_1}, \code{Y_Z_0_R_Z_0} (observed outcome if assigned control), \code{Y_Z_1_R_Z_1} (observed outcome if assigned treatment, "Missing" if attrited).
}
\keyword{datasets}
