% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_8_1}
\alias{table_8_1}
\title{Table 8.1 from FEDAI}
\format{
A data.frame
}
\usage{
table_8_1
}
\description{
Hypothetical potential outcomes for three employees depending on who is honored with an employee of the month award.
}
\keyword{datasets}
