% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_8_2}
\alias{table_8_2}
\title{Table 8.2 from FEDAI}
\format{
A data.frame
}
\usage{
table_8_2
}
\description{
Hypothetical potential outcomes for 5 villages depending on their treatment status and the treatment status of adjacent villages.
}
\keyword{datasets}
