% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_8_4}
\alias{table_8_4}
\title{Table 8.4 from FEDAI}
\format{
A data.frame
}
\usage{
table_8_4
}
\description{
Hypothetical potential outcomes for 30 policing hotspots locations depending on their treatment status and the treatment status of adjacent hotspots.
}
\keyword{datasets}
