% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_8_4_distmat}
\alias{table_8_4_distmat}
\title{Distance matrix for Table 8.4 from FEDAI}
\format{
A data.frame
}
\usage{
table_8_4_distmat
}
\description{
Distance matrix for the 30 hypothetical policing hotspots locations described in Table 8.4
}
\keyword{datasets}
