% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_8_7}
\alias{table_8_7}
\title{Table 8.7 from FEDAI}
\format{
A data.frame
}
\usage{
table_8_7
}
\description{
Hypothetical potential outcomes for 8 media markets over three weeks, depending on current and previous period treatment status.
}
\keyword{datasets}
