% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{table_9_2}
\alias{table_9_2}
\title{Table 9.2 from FEDAI}
\format{
A data.frame
}
\usage{
table_9_2
}
\description{
Tidy aggregate reply rates from Rosen (2010) stratified by legislator ethnicity. Columns: \code{X_legislator_hispanic} (TRUE/FALSE), \code{Z_author}, \code{Z_good_grammar}, \code{Y_reply_pct}, \code{N}.
}
\keyword{datasets}
