% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titiunik_2010}
\alias{titiunik_2010}
\title{titiunik_2010 from FEDAI}
\format{
A data.frame
}
\usage{
titiunik_2010
}
\description{
Titiunik, R. (2010). Exercise dataset for FEDAI Chapter 13. Regression discontinuity dataset on term lengths and legislative productivity.
Columns: \code{year} (legislative year), \code{party} (party affiliation), \code{dshort_term} (short-term indicator), \code{votesh_incumb} (incumbent vote share), \code{usrepvotesh_dem} (US representative Democratic vote share), \code{bills_intro} (bills introduced).
}
\keyword{datasets}
